spring-guides/gs-multi-module/library/src/main/java/io/lenur/multimodule/service/ApplicationService.java
package io.lenur.multimodule.service;

import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Service;

@Service
@EnableConfigurationProperties(ServiceProperties.class)
public class ApplicationService {
    private final ServiceProperties serviceProperties;

    public ApplicationService(ServiceProperties serviceProperties) {
        this.serviceProperties = serviceProperties;
    }

    public String message() {
        return this.serviceProperties.getMessage();
    }
}